Addded -Wconversion -Wpointer-arith flags for host build #175

+++ host/ocalls.c
@@ -77,7 +77,7 @@ void HandleThreadWait(oe_enclave_t* enclave, uint64_t arg_in)
 
 #if defined(__linux__)
 
-    if (__sync_fetch_and_add(&event->value, -1) == 0)
+    if (__sync_fetch_and_add(&event->value, (uint32_t)-1) == 0)
     {
         do
         {
@@ -214,7 +214,8 @@ static char** _backtrace_symbols(
     /* Determine total memory requirements */
     {
         /* Calculate space for the array of string pointers */
-        if (oe_safe_mul_sizet(size, sizeof(char*), &malloc_size) != OE_OK)
+        if (oe_safe_mul_sizet((size_t)size, sizeof(char*), &malloc_size) !=
+            OE_OK)
             goto done;
 
         /* Calculate space for each string */
@@ -244,7 +245,7 @@ static char** _backtrace_symbols(
     ret = (char**)ptr;
 
     /* Skip over array of strings */
-    ptr += size * sizeof(char*);
+    ptr += (size_t)size * sizeof(char*);
 
     /* Copy strings into return buffer */
     for (int i = 0; i < size; i++)
